[FIX] Address merge request
